refactor: move SearchPanel to Header
