+++ src/components/Header/components/SearchPanel/SearchPanel.scss
