+++ src/components/Header/components/SearchPanel/__tests__/searchPanel.test.jsx
