+++ src/components/Header/components/SearchPanel/assets/images/search-icon.png
