+++ src/components/Header/components/SearchPanel/index.js
